Refactor monolith into the tpp_solver package
Split the 2,200-line tpp_solver_mt.py into a focused package so the
computational core can be tested independently of Streamlit:

- normalization, models, preprocessing, database, io_utils, fitting,
  plotting, statistics, annotation and ui modules, plus a moved
  readme_content.
- tpp_solver_mt.py is now a thin entry point that calls tpp_solver.ui.main,
  preserving the `streamlit run tpp_solver_mt.py` command and Docker entry.
- Extract a pure benjamini_hochberg() FDR helper from the statistics UI
  function and cover it with tests.
- Remove seven dead functions (average_samples, get_replicant_lists,
  count_invalid_rows, save_as_svg, save_results_to_zip, create_results_zip,
  display_summary_tables).
- pyproject declares the package; tests import from the package modules
  and a new test_statistics suite is added (24 tests total).

Function bodies were moved verbatim; behavior is unchanged. Verified with
ruff, the full test suite, an import smoke test of the whole module graph,
and a multiprocessing run confirming the relocated worker pickles under the
spawn start method.
