+++ tpp_solver/__init__.py
@@ -1,0 +1,3 @@
+"""TPP Solver: Thermal Proteome Profiling analysis package."""
+
+__version__ = "1.2.0"
