+++ tpp_solver/readme_content.py
